git-svn-id: https://svn-dares-dart.cgd.ucar.edu/DART/branches/recam@12384 dfa8782c-da17-4c45-ba5c-5625b50a00d6

Former-commit-id: 1eaffdf094ef99d69846b1decf361cbb1132c82e

+++ models/lorenz_96_2scale/model_mod.f90
@@ -329,7 +329,7 @@ end subroutine adv_1step
 
 function get_model_size()
 
-integer :: get_model_size
+integer(i8) :: get_model_size
 
 get_model_size = l96%model_size
